+++ src/__tests__/getStringFromFile.test.ts
@@ -12,49 +12,51 @@ describe("read data XMLs", () => {
 	const dataXmlStrings: Array<string> = [];
 
 	test("get back string from utf8 xml fills", async () => {
-		const [error1, documentXmlString1] = useTry(() => {
+		const [error1, dataXmlString1] = useTry(() => {
 			return getStringFromFile(testUtf8File1);
 		});
-		const [error2, documentXmlString2] = useTry(() => {
+		const [error2, dataXmlString2] = useTry(() => {
 			return getStringFromFile(testUtf8File2);
 		});
 
 		expect(error1).toBeNull();
 		expect(error2).toBeNull();
 
-		dataXmlStrings.push(documentXmlString1);
-		dataXmlStrings.push(documentXmlString2);
+		dataXmlStrings.push(dataXmlString1);
+		dataXmlStrings.push(dataXmlString2);
 	});
 
 	test("get back string from utf16 xml fills", async () => {
-		const [error1, documentXmlString1] = useTry(() => {
+		const [error1, dataXmlString1] = useTry(() => {
 			return getStringFromFile(testUtf16File1);
 		});
-		const [error2, documentXmlString2] = useTry(() => {
+		const [error2, dataXmlString2] = useTry(() => {
 			return getStringFromFile(testUtf16File2);
 		});
 
 		expect(error1).toBeNull();
 		expect(error2).toBeNull();
 
-		dataXmlStrings.push(documentXmlString1);
-		dataXmlStrings.push(documentXmlString2);
+		dataXmlStrings.push(dataXmlString1);
+		dataXmlStrings.push(dataXmlString2);
 	});
 
 	test("utf16 and utf8 files return valid XML", async () => {
-		for (const documentXmlString of dataXmlStrings) {
-			expect(validate(documentXmlString)).toEqual(true);
+		for (const dataXmlString of dataXmlStrings) {
+			expect(validate(dataXmlString)).toEqual(true);
 		}
 	});
 
 	test("utf16 and utf8 files contain proper tags for data XML", async () => {
-		for (const documentXmlString of dataXmlStrings) {
-			const documentJson = parse(documentXmlString, {
+		for (const dataXmlString of dataXmlStrings) {
+			const dataJson = parse(dataXmlString, {
 				ignoreAttributes: false,
 				attributeNamePrefix: "",
 			});
 
-			expect(documentJson?.Documents?.items?.item == null).toEqual(false);
+			expect(dataJson[Object.keys(dataJson)[0]]?.items?.item == null).toEqual(
+				false
+			);
 		}
 	});
 });