+++ src/fake/frakeWriteFakeFiles.ts
@@ -1,23 +1,23 @@
 import fs from "fs-extra";
 import { v4 as uuidv4 } from "uuid";
-import { ChiliDocument } from "../common/types";
+import { ChiliItem } from "../common/types";
 
 export default function writeFakeFiles(
-	documents: Array<ChiliDocument>,
+	items: Array<ChiliItem>,
 	outputPath: string,
 	skipFiles: boolean
-): Array<ChiliDocument> {
+): Array<ChiliItem> {
 	const skippedFiles = [];
 
-	for (let j = 0; j < documents.length; j++) {
-		const document = documents[j];
+	for (let j = 0; j < items.length; j++) {
+		const item = items[j];
 
-		const path = outputPath + "\\" + document.relativePath;
+		const path = outputPath + "\\" + item.relativePath;
 
 		const skip = skipFiles ? Math.floor(Math.random() * 10) : 10;
 
 		if (skip <= 2) {
-			skippedFiles.push(document);
+			skippedFiles.push(item);
 			continue;
 		}